CORE-7177 Fix some package placing
