+++ components/ledger/ledger-utxo-flow/src/main/kotlin/net/corda/ledger/utxo/flow/impl/transaction/UtxoTransactionMetaData.kt
@@ -1,3 +1,3 @@
-package net.corda.ledger.utxo.impl.transaction
+package net.corda.ledger.utxo.flow.impl.transaction
 
 const val TRANSACTION_META_DATA_UTXO_LEDGER_VERSION = "0.001"
