+++ components/ledger/ledger-utxo-flow/src/main/kotlin/net/corda/ledger/utxo/flow/impl/transaction/factory/UtxoTransactionBuilderFactory.kt
@@ -1,4 +1,4 @@
-package net.corda.ledger.utxo.impl.transaction.factory
+package net.corda.ledger.utxo.flow.impl.transaction.factory
 
 import net.corda.v5.ledger.utxo.transaction.UtxoTransactionBuilder
