+++ testing/ledger/ledger-utxo-testkit/src/main/kotlin/net/corda/ledger/utxo/testkit/UtxoTimeWindowExample.kt
@@ -1,6 +1,6 @@
 package net.corda.ledger.utxo.testkit
 
-import net.corda.ledger.utxo.impl.timewindow.TimeWindowBetweenImpl
+import net.corda.ledger.utxo.flow.impl.timewindow.TimeWindowBetweenImpl
 import java.time.Instant
 
 val utxoTimeWindowExample = TimeWindowBetweenImpl(Instant.MIN, Instant.MAX)
